refactor: extract subcommand registration into per-area modules (2786 -> 328 lines) (#421)
* refactor(commands): register subcommands from per-area modules

commands.lua was 2786 lines with all 37 :Gitflow subcommands registered
inside one register_builtin_subcommands function, sitting at 59 of
LuaJIT's 60 upvalues. Eight queued panel features each had to edit that
one function, so they serialized.

Registration now lives in eight area modules under lua/gitflow/commands/,
listed in AREA_MODULES. An area exports register(ctx) and an optional
ordered panels list that :Gitflow close walks, so a new panel adds a
module plus one line instead of editing the monolith.

Two seams moved onto the subcommand spec:
- category, previously derived from name->category tables in commands.lua
- complete, previously a name-keyed if-chain in M.complete

Both yield identical output; register_subcommand now validates complete.
commands.lua is down to 328 lines of wiring and dispatch, and
register_builtin_subcommands from 59 upvalues to 3 (worst area: 20).

No behavior change: usage(), the subcommand key set, palette_entries,
completion for 52 cmdlines and dispatch messages for 52 arg vectors are
byte-identical before and after.

* test(stage2): enumerate modules for the confirm() accelerator scan

The scan read a hardcoded file list, so moving the two confirms out of
commands.lua into the workspace/github command areas dropped them from
coverage without failing anything.

Globbing lua/**/*.lua instead covers 16 modules with choice lists, up
from 9. The hardcoded list had already rotted well before this refactor:
reflog, reset, revert, review, tag and worktree panels were never
scanned, hiding 15 choice lists (7 in review.lua alone). No conflicts in
any of them, so this only widens coverage.

The commit-confirm check pointed at commands.lua and had gone vacuous for
the same reason; it now reads the workspace area and asserts the confirm
is actually there, so it fails loudly instead of silently passing if the
flow moves again.

---------
