+++ .dm-knowledge/gf-cmd-extract.md
@@ -1,0 +1,2 @@
+- **[convention]** :Gitflow subcommands register from per-area modules under lua/gitflow/commands/, listed in AREA_MODULES in commands.lua. An area exports register(ctx) -- ctx.register(name, spec), ctx.config, ctx.commands -- and an optional ordered panels list that :Gitflow close walks. A spec is {description, run(cmd), category?, complete?}; commands.lua is wiring+dispatch only and never holds registrations. Adding a panel = new area module + one AREA_MODULES line.  _(2026-07-20T14:16:44Z)_
+- **[pitfall]** The 60-upvalue ceiling on commands.lua's register_builtin_subcommands is gone: it dropped from 59 to 3 upvalues once registrations moved to per-area modules (worst area is 20). Supersedes the gf-commands-bugs.md note that says to avoid new file-level locals there.  _(2026-07-20T14:16:44Z)_
